feat:齐套天津百件单位判断

+++ mes-modules/produce/src/main/java/com/richfit/mes/produce/entity/extend/ProjectBomComplete.java
@@ -27,17 +27,17 @@ public class ProjectBomComplete extends ProjectBom {
     /**
      * 已安装数量
      */
-    private int installNumber;
+    private double installNumber;
     /**
      * 第三方仓储数量
      */
-    private int erpNumber;
+    private double erpNumber;
     /**
      * 本地仓储数量
      */
-    private int storeNumber;
+    private double storeNumber;
     /**
      * 缺件数量
      */
-    private int missingNumber;
+    private double missingNumber;
 }
